Modified UpdateP to give Paetkau-like results
I changed a few lines in UpdateP so that the
Dirichlet parameters are all very large (low variance)
but proportional to the counts, except for absent
alleles which get a Dirichlet parameter that is
.01 * total_number_of_gene_copies * big_num
where big_num is the amount I increase all the counts by.

The result of this is that the updated allele frequency
is nearly identical to what you would get if you just
counted up all the alleles in each population and then
normalized them to sum to one, and then you added 0.01
to each population at each allele which did not appear
in that population (essentially the Paetkau-like
correction).

I included a test file that runs the test data through
with and without freqscorr (my hack only gets executed
with Freqscorr = 0) and shows that the allele freqs are
pretty similar.

You run that test by doing:
```sh
./test/test_alle_freq_hack.sh
```