Pretty much done with assignment mods.
Still need to integrate the allele frequency code
branch.  But what I have done here is:

1. Made some global variables to count itertion numbers and such.
2. I let BURNIN be the number of flock iterations and it just keeps
doing them until it is out of NUMREPS.  So the combination of BURNIN
and NUMREPS determines the number of iterations per flock rep and
the number of flock reps.
3. After each flock rep I run the initialization function to start anew.
4. Currently the output is tagged and spit out to stdout.
5. I need to put headers on for the different types of tagged output.
6. In case the allele frequency variation (even with my alle freq update hacks) is too great to assess partitions on the basis of the
log likelihood I print out a FLOCKTURE_CLUSTER_STRING that can be
used to define sets of individuals and then you can table of unique
partitions easily in R.
7. I had to set ADMBURNIN in extraparams to -1.

So, basically, I just need to put some output headers on and then
merge this branch into the allele freq mods branch, and I'll have
flockture.  Then we can drive it from within R.  Cool.